requirements.txt
streamlit>=1.44.0
openai>=1.75.0
python-dotenv>=1.0.1
reportlab>=4.2.0
python-docx>=1.1.2
pypdf>=5.4.0
supabase>=2.7.4
openpyxl>=3.1.5
